# docs/openapi/cliser-data-node.openapi.current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CLISER DATA NODE API",
    "version": "1.0.0",
    "description": "HTTP API do CLISER DATA NODE.\n\nAPI Version: v1\nBase Path: /api/v1"
  },
  "paths": {
...
    "/api/v1/namespaces/{namespace}/uploads": {
      "post": {
        "tags": [
          "uploads"
        ],
        "summary": "Upload Object",
        "operationId": "upload_object_api_v1_namespaces__namespace__uploads_post",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Namespace"
            },
            "name": "namespace",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_object_api_v1_namespaces__namespace__uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                },
                "examples": {
                  "UploadResponse": {
                    "$ref": "#/components/examples/UploadResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_v1_transactions_get",
        "parameters": [
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponse"
                },
                "examples": {
                  "TransactionListResponse": {
                    "$ref": "#/components/examples/TransactionListResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Get Transaction",
        "operationId": "get_transaction_api_v1_transactions__transaction_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Id"
            },
            "name": "transaction_id",
            "in": "path"
          },
          {
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            },
            "name": "authorization",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                },
                "examples": {
                  "TransactionResponse": {
                    "$ref": "#/components/examples/TransactionResponse"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                },
                "examples": {
                  "ErrorValidation": {
                    "$ref": "#/components/examples/ErrorValidation"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_upload_object_api_v1_namespaces__namespace__uploads_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_object_api_v1_namespaces__namespace__uploads_post"
      },
...
      "TransactionListResponse": {
        "properties": {
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Transactions"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "transactions",
          "count"
        ],
        "title": "TransactionListResponse"
      },
      "TransactionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "operation",
          "state",
          "created_at",
          "updated_at"
        ],
        "title": "TransactionResponse"
      },
      "UploadResponse": {
        "properties": {
          "object_id": {
            "type": "string",
            "title": "Object Id"
          },
          "namespace": {
            "type": "string",
            "title": "Namespace"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "object_id",
          "namespace",
          "filename",
          "size",
          "content_hash",
          "status"
        ],
        "title": "UploadResponse"
      },
...
      "APIError": {
        "title": "APIError",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "OBJECT_NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Objeto não encontrado"
              },
              "request_id": {
                "type": [
                  "string",
                  "null"
                ],
                "example": "req_01JEXAMPLE"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "credential",
        "description": "Credencial de API enviada no cabeçalho Authorization como Bearer <credential>."
      }
    },
    "examples": {
...
      "UploadResponse": {
        "value": {
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "filename": "documento.pdf",
          "size": 245760,
          "content_hash": "sha256:01JEXAMPLE",
          "status": "ACTIVE"
        },
        "summary": "Upload concluído"
      },
...
      "TransactionResponse": {
        "value": {
          "transaction_id": "tx_01JEXAMPLE",
          "object_id": "obj_01JEXAMPLE",
          "namespace": "documents",
          "operation": "OBJECT_CREATE",
          "state": "COMMITTED",
          "created_at": "2026-09-15T15:00:00Z",
          "updated_at": "2026-09-15T15:00:01Z"
        },
        "summary": "Transação concluída"
      },
      "TransactionListResponse": {
        "value": {
          "transactions": [
            {
              "transaction_id": "tx_01JEXAMPLE",
              "object_id": "obj_01JEXAMPLE",
              "namespace": "documents",
              "operation": "OBJECT_CREATE",
              "state": "COMMITTED",
              "created_at": "2026-09-15T15:00:00Z",
              "updated_at": "2026-09-15T15:00:01Z"
            }
          ],
          "count": 1
        },
        "summary": "Lista de transações"
      },
...
      "ErrorValidation": {
        "value": {
          "error": {
            "code": "VALIDATION_ERROR",
            "message": "Request validation failed.",
            "request_id": "req_01JEXAMPLE",
            "details": {
              "errors": [
                {
                  "type": "value_error",
                  "loc": [
                    "body",
                    "namespace"
                  ],
                  "msg": "field required"
                }
              ]
            }
          }
        },
        "summary": "Erro de validação"
      },
...
    }
  },
...
}